+++ src/index/blockfilterindex.cpp
@@ -8,6 +8,7 @@
 #include <hash.h>
 #include <index/blockfilterindex.h>
 #include <node/blockstorage.h>
+#include <util/fs_helpers.h>
 #include <util/system.h>
 #include <validation.h>
