+++ src/init/common.cpp
@@ -11,6 +11,7 @@
 #include <logging.h>
 #include <node/interface_ui.h>
 #include <tinyformat.h>
+#include <util/fs_helpers.h>
 #include <util/string.h>
 #include <util/system.h>
 #include <util/time.h>
