+++ src/kernel/mempool_persist.cpp
@@ -15,7 +15,7 @@
 #include <sync.h>
 #include <txmempool.h>
 #include <uint256.h>
-#include <util/system.h>
+#include <util/fs_helpers.h>
 #include <util/time.h>
 #include <validation.h>
